@@ -1,24 +1,21 @@
 // IMatchFinder.cs
 
-using System;
+namespace SevenZip.Compression.LZ;
 
-namespace SevenZip.Compression.LZ
+interface IInWindowStream
 {
-	interface IInWindowStream
-	{
-		void SetStream(System.IO.Stream inStream);
-		void Init();
-		void ReleaseStream();
-		Byte GetIndexByte(Int32 index);
-		UInt32 GetMatchLen(Int32 index, UInt32 distance, UInt32 limit);
-		UInt32 GetNumAvailableBytes();
-	}
+    void SetStream(System.IO.Stream inStream);
+    void Init();
+    void ReleaseStream();
+    byte GetIndexByte(int index);
+    uint GetMatchLen(int index, uint distance, uint limit);
+    uint GetNumAvailableBytes();
+}
 
-	interface IMatchFinder : IInWindowStream
-	{
-		void Create(UInt32 historySize, UInt32 keepAddBufferBefore,
-				UInt32 matchMaxLen, UInt32 keepAddBufferAfter);
-		UInt32 GetMatches(UInt32[] distances);
-		void Skip(UInt32 num);
-	}
+interface IMatchFinder : IInWindowStream
+{
+    void Create(uint historySize, uint keepAddBufferBefore,
+            uint matchMaxLen, uint keepAddBufferAfter);
+    uint GetMatches(uint[] distances);
+    void Skip(uint num);
 }
